refactor(rust): extract document mutations (slice 8/9)
Move document mutation and import operations into db/documents.rs:

- json_to_bson_document
- delete_document_impl, delete_many_impl
- update_many_impl, insert_document_impl, update_document_impl
- import_documents_impl and ImportResult

Keep crate-root re-exports so command wrappers and tests continue to use the same symbols. Behavior unchanged.

cargo check: passed
cargo test: 60 passed
