+++ src-tauri/src/db/mod.rs
@@ -1,6 +1,7 @@
 //! Database-operation modules (split out of lib.rs).
 
 pub mod aggregate;
+pub mod documents;
 pub mod export;
 pub mod gridfs;
 pub mod metadata;
